kochen eingebaut

+++ src/main/java/com/schmudlach/lagersystem/dto/VerfügbareLebensmittelDTO.java
@@ -2,7 +2,7 @@ package com.schmudlach.lagersystem.dto;
 
 public record VerfügbareLebensmittelDTO(
         int lebensmittelId,
-        int anzahl,
+        double anzahl,
         int threshold
 ) {
 }
